10-carpentries: Remind learners that motivating PRs is important
Try to head off PR and commit-message summaries that don't tell the
lesson maintainer what the change is about, or why the submitter
thinks it's a good idea.

I've put an abbreviated form of this comment in the training-checkout
reminder, with an fully qualified link to more detailed 10-carpentries
entry.  The messages/ entries look like email templates, neither a
relative link (which wouldn't resolve) or a detailed description
(which would be distracting) seem like good fits there.  If the
messages/ entries are email templates, I'd suggest injecting some YAML
front matter [1] so we can use:

  {{ site.domain }}{{ site.baseurl }}/10-carpentries.html#final-steps

in that link.  But that seemed like a bigger change than I should be
biting off in one commit.

[1]: http://jekyllrb.com/docs/frontmatter/