Removes content in _extras/discuss.md.
 This content appears to be a laundry list of things that were cut from the
episodes previously and much was outdated (e.g. we **do** now have a
systematic assessment program).

If there's anything that was deleted here that folks think should stay,
please make the case in an issue. If there is content that we should keep,
I'd like to integrate it into lessons as the discuss file is probably not
widely read.