Adding learning objectives in new form
This is a suggestion to retain some of the learning objectives material, but focuses that content more specifically in a way that instructors can use when preparing to teach a lesson as opposed to designing one. Also modifies some language re challenge question design to address redundancy with existing material and to encourage instructors to engage in design even where challenges are already present.