Clarify & shorten activity
This activity had an indefinite run time and it was awkward to choose which questions to address now and which not to. With live coding now coming up next, one or both instructors can sit for a few minutes to write responses in the etherpad while learners practice. Will insert a prompt to return to this later once most questions have been addressed by the day's curriculum.