add DC and SWC mission
Added DC and SWC missions as per #785 . This doesn't add a a challenge exercise as suggested, but it does provide a little more meat in the opening paragraph. 

One thing to consider is rewriting this whole episode. I think the venn diagram and some of the lists could be condensed into 1 table, for example.