Change activity order in Management
Existing order creates a major problem with CoC discussion last -- it ends up crunched for time and leaves everybody on edge as they head off to lunch. This PR moves management challenges and CoC up to come before co-instruction. This allows both of these topics to carry forward into conversations about co-instruction, and leaves an easier choice for cutting content to get to lunch on time.