Remove self-fulfilling  prophecy
The section I changed represents a self-fulfilling itself. If you tell scientist they don't want to code, they likely don't want to code. It is a bit of stereo-type threat and self-fulfilling prophecy. Hence, I changed it to make it more inviting and I think by know also more true. There are lots of scientists out there that want to code but don't know how to start and where. I think this is also illustrated by the success of the carpentry.