replaced phrase on errors in ep 8 line 36
I suggest to replace "when done well" with "when handled transparently" in Line 36. Done well sounds like performance-based feedback, and is generic in nature. I understand the point of this section is to normalise and make space for errors in the workshop, which is done by handling errors transparently and honestly.