Loosen activity prompt
This might help to make it clear that learners are to draw upon their own intuition about what is 'basic' or 'advanced' for the activity. Still does not solve the problem if they feel that they lack intuition, but at least should support them in engaging in guesswork rather than wondering if an explanation was missing.