More accurately described the expertise-reversal effect
The existing discussion of the expertise-reversal effect was inaccurate. The effect does not suggest that experts are "less good" at teaching than novices. What it suggests is that the relative amount of instructional support and guidance has differential impact depending on the level of expertise in a topic area. Basically, it means that for experts, walking through all the details of concepts actually can have a negative impact on performance; while for novices, the absence of these materials can have a negative impact.

For example:

"The expertise reversal is a reversal in the relative effectiveness of instructional methods as levels of learner knowledge in a domain change." - https://link.springer.com/article/10.1007/s11251-009-9102-0

https://impact.chartered.college/article/expertise-reversal-effect-instructional-implications/

I've tried to remain faithful to the original points raised, but be scientifically accurate at the same time. I have also removed the unfounded notion that "experts" are poor lecturers, which there is little evidence to substantiate this claim.