Replace "instructor-of-record" with better understandable term
I suggest to replace "instructor-of-record" with main/responsible teacher, or a similar, more universally understandable term.
At our recent instructor training, no-one knew what an "instructor of record" was. We discussed this briefly and thought that person would not have much responsibility, and when someone checked we were surprised to find out that this is the main responsible teacher. This seems to be a term that is only used in some parts of the world (US, maybe?). Participants of the training were mainly from Europe, Middle East, India/Pakistan and Africa.