Revise whole people wording
Original "whole people" (line 64) was clunky/confusing but suggested revision "people as a whole" could be taken to mean the overall makeup of the group as opposed to what was I think intended, the various aspects of the full self of each person. In light of the following paragraph, I think individual avoids ambiguity and suffices to convey the message.