Replace cdh with Collaborative Lesson Development Training
Replace link to the retired curriculum development handbook (which states):
`The Carpentries is preparing to retire this handbook at the end of 2023. Existing content that has not already been replicated elsewhere will be relocated. See the relevant issue on the source repository for more details.`

Add link to current training.