remove dismissive language
The curriculum offers some resources about using pronouns.  The language "For those unfamiliar with..." reminds some users that there is a concept they are unfamiliar with, and does not seem to have any other purpose.  

I am recommending removing that language and only offering the resources.  If we are offering the resources, we should offer them as they are, without contextualizing that some people don't know about them.